Aggiunta gestione semantic lock microservizio utenti

+++ User-Service/src/main/java/org/caesar/userservice/Data/Services/Impl/UserCardServiceImpl.java
@@ -34,6 +34,10 @@ public class UserCardServiceImpl implements UserCardService {
 
         //Presa della lista delle carte associate all'utente
         UserCard userCard = userCardRepository.findById(id).orElse(null);
+
+        if(userCard == null || userCard.isOnDeleting())
+            return null;
+
         return modelMapper.map(userCard, UserCardDTO.class);
     }
 
@@ -46,6 +50,9 @@ public class UserCardServiceImpl implements UserCardService {
         List<UUID> result = new Vector<>();
 
         for (UserCard userCard: userCards) {
+            if(userCard.isOnDeleting())
+                return null;
+
             result.add(userCard.getId());
         }
 
@@ -59,6 +66,8 @@ public class UserCardServiceImpl implements UserCardService {
         List<UserCard> userCards = userCardRepository.findAllByUserUsername(userUsername);
 
         for(UserCard ut: userCards) {
+            if(ut.isOnDeleting())
+                return null;
             result.add(modelMapper.map(ut, UserCardDTO.class));
         }
 
@@ -69,7 +78,7 @@ public class UserCardServiceImpl implements UserCardService {
     public boolean checkCard(String username, UUID cardId) {
         UserCard userCard= userCardRepository.findByUserUsernameAndId(username, cardId);
 
-        return userCard != null;
+        return userCard != null && !userCard.isOnDeleting();
     }
 
     //Aggiunta della relazione carta utente
@@ -78,6 +87,7 @@ public class UserCardServiceImpl implements UserCardService {
         //Try per gestire l'errore nell'inserimento della tupla (l'eventuale rollback sarà gestito dal @Transactional del save()
         try{
             UserCard userCardEntity = modelMapper.map(userCard, UserCard.class);
+            userCardEntity.setOnDeleting(false);
             userCardRepository.save(userCardEntity);
 
             return true;
@@ -91,6 +101,11 @@ public class UserCardServiceImpl implements UserCardService {
     @Override
     public boolean deleteUserCard(UserCardDTO userCardDTO) {
         try {
+            UserCard userCard= userCardRepository.findById(userCardDTO.getId()).orElse(null);
+
+            if(userCard == null || userCard.isOnDeleting())
+                return false;
+
             userCardRepository.deleteById(userCardDTO.getId());
             return true;
         } catch (Exception | Error e) {
@@ -109,6 +124,9 @@ public class UserCardServiceImpl implements UserCardService {
 
             if(cards.isEmpty())
                 return new Vector<>();
+
+            if(cards.getFirst().isOnDeleting() && !rollback)
+                return null;
 
             List<CardDTO> result= new Vector<>();
             for(UserCard userCard: cards) {
